payments flattened
